split scf gradient algorithm files
